X-Y plot option (#9)
Add an X-Y plot option from a signal table mixin. X-Y ordering is
specified by the order in which the signals are selected. X-Y line fades
with older time / smaller index. Synchronized with regions, if active.
Supports transforms.

Also fixes a datatype issue with the example waveform and improves
transform error messages.

Adds a few base mixins HasRegionsSignalsTable and HasDataSignalsTable to
deduplicate some shared code.

+++ pyqtgraph_scope_plots/csv/__main__.py
@@ -47,7 +47,7 @@ if __name__ == "__main__":
     plots._set_data(
         {
             "sine": (xs, np.sin(xs / X_PER_CYCLE * 2 * math.pi)),
-            "square": (xs, np.signbit(np.sin(xs / X_PER_CYCLE * 2 * math.pi))),
+            "square": (xs, np.signbit(np.sin(xs / X_PER_CYCLE * 2 * math.pi)).astype(float)),
             "cycle": (
                 xs,
                 [f"Cycle {int(x)}" for x in np.floor(xs / X_PER_CYCLE)],
